Build sunxi_emac as a module
In order to support emac and gmac in the same fex file,
where they are both enabled, build emac and gamc both
as modules and load them when needed in /etc/modules.

When emac enabled in fex, and built as built-in in kernel,
gmac won't work.

When emac and gamc both enabled in fex, and built both as
kernel modules, then we can load them as necessary in
different setup as in the a20-som-evb and a20-olinuxino-micro